# generated/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
...
      "IHttpErrorDto": {
        "properties": {
          "name": {
            "type": "string"
          },
          "errorKey": {
            "type": "string"
          },
          "validationErrors": {
            "items": {
              "$ref": "#/components/schemas/IApiValidationError"
            },
            "type": "array"
          },
          "cause": {
            "$ref": "#/components/schemas/TErrorCause"
          }
        },
        "required": ["name", "errorKey"],
        "type": "object",
        "additionalProperties": false
      },
...
      "UPaginationSortDirection": {
        "type": "string",
        "enum": ["asc", "desc"]
      },
...
      "IJournalEntryListDto": {
        "properties": {
          "sourceType": {
            "$ref": "#/components/schemas/UJournalEntrySourceType"
          },
          "memo": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/UJournalEntryStatus"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time"
          },
          "postedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "voidedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "voidingEntryId": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "number",
            "format": "double"
          },
          "createdBy": {
            "$ref": "#/components/schemas/TEntityId"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "accountingEntityId": {
            "type": "string"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/IFileAttachment"
            },
            "type": "array"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/IJournalLineListDto"
            },
            "type": "array"
          }
        },
        "required": [
          "sourceType",
          "memo",
          "status",
          "effectiveDate",
          "postedAt",
          "voidedAt",
          "voidingEntryId",
          "version",
          "createdBy",
          "createdAt",
          "updatedAt",
          "id",
          "accountingEntityId",
          "attachments",
          "lines"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "IPaginatedResponse_IJournalEntryListDto_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/IJournalEntryListDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/IPaginationResponseMeta"
          }
        },
        "required": ["data", "meta"],
        "type": "object",
        "additionalProperties": false
      },
      "UJournalEntrySortBy": {
        "type": "string",
        "enum": ["createdAt", "effectiveDate"]
      },
...
      "IJournalEntryDto": {
        "properties": {
          "sourceType": {
            "$ref": "#/components/schemas/UJournalEntrySourceType"
          },
          "memo": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/UJournalEntryStatus"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time"
          },
          "postedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "voidedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "voidingEntryId": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "number",
            "format": "double"
          },
          "createdBy": {
            "$ref": "#/components/schemas/TEntityId"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "accountingEntityId": {
            "type": "string"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/IFileAttachment"
            },
            "type": "array"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/IJournalLineDto"
            },
            "type": "array"
          }
        },
        "required": [
          "sourceType",
          "memo",
          "status",
          "effectiveDate",
          "postedAt",
          "voidedAt",
          "voidingEntryId",
          "version",
          "createdBy",
          "createdAt",
          "updatedAt",
          "id",
          "accountingEntityId",
          "attachments",
          "lines"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "IPaymentEntryReq": {
        "properties": {
          "attachmentReferences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "description": "Opaque handles returned when preparing file uploads. Each corresponding\nfile must be uploaded before the payment is created."
          },
          "sourceLine": {
            "$ref": "#/components/schemas/IPaymentEntryLineReq"
          },
          "destinationLines": {
            "items": {
              "$ref": "#/components/schemas/IPaymentEntryLineReq"
            },
            "type": "array"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time"
          },
          "postedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "memo": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "sourceLine",
          "destinationLines",
          "effectiveDate",
          "postedAt",
          "memo"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "IReceiptEntryReq": {
        "properties": {
          "attachmentReferences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "description": "Opaque handles returned when preparing file uploads. Each corresponding\nfile must be uploaded before the receipt is created."
          },
          "sourceLines": {
            "items": {
              "$ref": "#/components/schemas/IReceiptEntryLineReq"
            },
            "type": "array"
          },
          "destinationLine": {
            "$ref": "#/components/schemas/IReceiptEntryLineReq"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time"
          },
          "postedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "memo": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "sourceLines",
          "destinationLine",
          "effectiveDate",
          "postedAt",
          "memo"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "IJournalEntryArchiveReq": {
        "properties": {
          "expectedVersion": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["expectedVersion"],
        "type": "object",
        "additionalProperties": false
      },
      "IJournalEntryDeletionReq": {
        "properties": {
          "expectedVersion": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["expectedVersion"],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "drimsheet-core",
    "version": "0.1.0-alpha.1",
    "description": "This repo contains the server-side source code for Drimsheet's bookkeeping and tax app.",
    "contact": {
      "name": "Osahon Oboite"
    }
  },
  "paths": {
...
    "/journal-entries": {
      "get": {
        "operationId": "getJournalEntries",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPaginatedResponse_IJournalEntryListDto_"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Get paginated journal entries with optional line participation filters",
        "tags": ["Journal Entry"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "orderBy",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortDirection",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UPaginationSortDirection"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "accountId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "counterpartyId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "orderBy",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UJournalEntrySortBy"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["posted", "archived"]
            }
          }
        ]
      }
    },
    "/journal-entries/{id}": {
      "get": {
        "operationId": "getJournalEntry",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IJournalEntryListDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Get a journal entry by id",
        "tags": ["Journal Entry"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "operationId": "deleteJournalEntry",
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Delete a never-posted journal entry or reverse a previously-posted entry.",
        "tags": ["Journal Entry"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IJournalEntryDeletionReq"
              }
            }
          }
        }
      }
    },
    "/journal-entries/payment": {
      "post": {
        "operationId": "createPayment",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IJournalEntryDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create payment journal entry",
        "tags": ["Journal Entry"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IPaymentEntryReq"
              }
            }
          }
        }
      }
    },
    "/journal-entries/receipt": {
      "post": {
        "operationId": "createReceipt",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IJournalEntryDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create receipt journal entry",
        "tags": ["Journal Entry"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IReceiptEntryReq"
              }
            }
          }
        }
      }
    },
    "/journal-entries/transfer": {
      "post": {
        "operationId": "createTransfer",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IJournalEntryDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create transfer journal entry",
        "tags": ["Journal Entry"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ITransferEntryReq"
              }
            }
          }
        }
      }
    },
    "/journal-entries/{id}/rectify": {
      "post": {
        "operationId": "rectifyJournalEntry",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IJournalEntryRectificationDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Correct a journal entry while preserving its accounting audit trail.",
        "tags": ["Journal Entry"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TJournalEntryRectificationReq"
              }
            }
          }
        }
      }
    },
    "/journal-entries/{id}/archive": {
      "post": {
        "operationId": "archiveJournalEntry",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IJournalEntryDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Archive a journal entry without changing its financial effect.",
        "tags": ["Journal Entry"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IJournalEntryArchiveReq"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}